restructure models
