+++ app/__init__.py
@@ -1,4 +1,5 @@
 from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
 
 # Initialize the app
 app = Flask(__name__, instance_relative_config=True)
@@ -8,3 +9,8 @@ from app import views
 
 # Load the config file
 app.config.from_object('config')
+
+db = SQLAlchemy(app)
+from app.models import User, Notice
+db.create_all()
+
