+++ app/models/__init__.py
@@ -1,0 +1,4 @@
+
+from .User import User
+from .Notice import Notice
+
