+++ config.py
@@ -2,3 +2,6 @@
 
 # Enable Flask's debugging features. Should be False in production
 DEBUG = True
+
+# SQLALCHEMY_DATABASE_URI = 'postgresql://localhost/whu'
+SQLALCHEMY_DATABASE_URI = 'sqlite:///./test2.db'
